+++ netket/operator/__init__.py
@@ -15,12 +15,13 @@
 from ._abstract_operator import AbstractOperator
 
 from ._discrete_operator import DiscreteOperator
+from ._discrete_operator_jax import DiscreteJaxOperator
 from ._local_operator import LocalOperator
 from ._graph_operator import GraphOperator
 from ._pauli_strings import PauliStrings
 from ._lazy import Adjoint, Transpose, Squared
 from ._heisenberg import Heisenberg
-from ._ising import Ising
+from ._ising import Ising, IsingJax
 from ._bose_hubbard import BoseHubbard
 
 from ._abstract_super_operator import AbstractSuperOperator