+++ test/operator/test_operator.py
@@ -1,6 +1,7 @@
 import netket as nk
 import numpy as np
 import netket.experimental as nkx
+from netket.operator import DiscreteJaxOperator
 
 import pytest
 import jax
@@ -11,6 +12,7 @@ operators = {}
 g = nk.graph.Hypercube(length=10, n_dim=1, pbc=True)
 hi = nk.hilbert.Spin(s=0.5, N=g.n_nodes)
 operators["Ising 1D"] = nk.operator.Ising(hi, g, h=1.321)
+operators["Ising 1D Jax"] = nk.operator.IsingJax(hi, g, h=1.321)
 
 # Heisenberg 1D
 g = nk.graph.Hypercube(length=10, n_dim=1, pbc=True)
@@ -113,6 +115,16 @@ for name, op in operators.items():
     if hi.is_finite and hi.n_states < 2**14:
         op_finite_size[name] = op
 
+operators_numba = {}
+for name, op in operators.items():
+    if not isinstance(op, DiscreteJaxOperator):
+        operators_numba[name] = op
+
+op_jax_compatible = {}
+for name, op in op_finite_size.items():
+    if hasattr(op, "to_jax_operator"):
+        op_jax_compatible[name] = op
+
 
 @pytest.mark.parametrize("attr", ["get_conn", "get_conn_padded"])
 @pytest.mark.parametrize(
@@ -206,7 +218,7 @@ def test_repr(op):
 
 
 @pytest.mark.parametrize(
-    "op", [pytest.param(op, id=name) for name, op in operators.items()]
+    "op", [pytest.param(op, id=name) for name, op in operators_numba.items()]
 )
 def test_get_conn_numpy_closure(op):
     hi = op.hilbert
@@ -237,10 +249,7 @@ def test_get_conn_numpy_closure(op):
         pytest.param(s, id=f"shape={s}")
         for s in [
             (2,),
-            (
-                2,
-                1,
-            ),
+            (2, 1),
             (2, 1, 1),
         ]
     ],
@@ -317,3 +326,47 @@ def test_operator_on_subspace():
 
     h4 = nk.operator.Heisenberg(hi, g, acting_on_subspace=0)
     assert h4.acting_on == h1.acting_on
+
+
+@pytest.mark.parametrize(
+    "op", [pytest.param(op, id=name) for name, op in op_jax_compatible.items()]
+)
+def test_operator_jax_conversion(op):
+    op_jax = op.to_jax_operator()
+    op_numba = op_jax.to_numba_operator()
+
+    # check round_tripping
+    np.testing.assert_allclose(op_numba.to_dense(), op.to_dense())
+
+    np.testing.assert_allclose(op_numba.to_dense(), op_jax.to_dense())
+
+    # test packing unpacking
+    data, structure = jax.tree_util.tree_flatten(op_jax)
+    op_jax2 = jax.tree_util.tree_unflatten(structure, data)
+    op_numba2 = op_jax2.to_numba_operator()
+    np.testing.assert_allclose(op_numba2.to_dense(), op.to_dense())
+
+    # check that it is hash stable
+    _, structure2 = jax.tree_util.tree_flatten(op.to_jax_operator())
+    assert hash(structure) == hash(structure2)
+    assert structure == structure2
+
+
+@pytest.mark.parametrize(
+    "op", [pytest.param(op, id=name) for name, op in op_jax_compatible.items()]
+)
+def test_operator_jax_getconn(op):
+    """Check that get_conn returns the same result for jax and numba operators"""
+    op_jax = op.to_jax_operator()
+
+    all_states = op.hilbert.all_states()
+
+    @jax.jit
+    def _get_conn_padded(op, s):
+        return op_jax.get_conn_padded(s)
+
+    sp, mels = op.get_conn_padded(all_states)
+    sp_j, mels_j = _get_conn_padded(op_jax, all_states)
+
+    np.testing.assert_allclose(sp, sp_j)
+    np.testing.assert_allclose(mels, mels_j)