+++ CHANGELOG.md
@@ -7,8 +7,10 @@
 This release requires Python 3.8 and Jax 0.4.
 
 ### New Features
-* `netket.callbacks.EarlyStopping` now supports relative tolerances for determining when to stop [#14XX](https://github.com/netket/netket/pull/1481).
+* `netket.callbacks.EarlyStopping` now supports relative tolerances for determining when to stop [#1481](https://github.com/netket/netket/pull/1481).
 * `netket.callbacks.ConvergenceStopping` has been added, which can stop a driver when the loss function reaches a certain threshold [#14XX](https://github.com/netket/netket/pull/1481).
+* A new base class {class}`netket.operator.DiscreteJaxOperator` has been added, which will be used as a base class for a set of operators that are jax-compatible [#1506](https://github.com/netket/netket/pull/1506).
+* {class}`netket.operator.IsingJax` is a new operator that behaves as `netket.operator.Ising` but is Jax-compatible, meaning that it can be used inside of jax-jitted contexts and works better with chunking. It can also be constructed starting from a standard Ising operator by calling `operator.to_jax_operator()` [#1506](https://github.com/netket/netket/pull/1506).
 
 ### Breaking Changes
 * The first two axes in the output of the samplers have been swapped, samples are now of shape `(n_chains, n_samples_per_chain, ...)` consistent with `netket.stats.statistics`. Custom samplers need to be updated to return arrays of shape `(n_chains, n_samples_per_chain, ...)` instead of `(n_samples_per_chain, n_chains, ...)`. [#1502](https://github.com/netket/netket/pull/1502)